Changes in Branch Course Master

+++ app/src/main/java/com/example/genius/ui/Homework_Fragment/HomeWorkCheckingFragment.java
@@ -63,7 +63,7 @@ public class HomeWorkCheckingFragment extends Fragment {
         {
             if (bundle.containsKey("ID"))
             {
-                id.setText(""+bundle.getInt("ID"));
+                id.setText(""+bundle.getLong("ID"));
             }
         }
